Split entirely the Connector code from Parser.
This is done in a new ``formats`` package, where old code has been copied
and briefly adapted to the new structure. The ``models`` modules are now
much lighter, because they contain just ModelConnector class definitions.

More changes are needed to complete a clear and proper transition to the
new FormatParser API, obviously.