[scripts] #18 #20 Display downloaded data in real-time.
The UI for downloading was changed a bit, it takes three clicks until
the end (two clicks before), but it is more responsive: dialogs won't
"sleep" after the user has clicked.

A more sophisticated version of this UI could add a status bar, so
that there is no "waiting for data" in the text area, and we could even
show the total amount of bytes downloaded.