[core] fix #27 refactor the models subpackage entirely.
The list of builtin models is now treated in the same way as input
and output formats. Custom connection is now only slightly different,
accepting generic Serial parameters, and it is using exactly the same
interface as other model Connectors.

This is a first iteration: the next step is to just define the needed
connection parameters in a dict, and let users add more models, that will
be saved in a JSON file along with the user preferences.