Backout 30766de
To my best understanding, the ?corrections? are already included in the
proper data and the comment lines are there just for reference, not for
being parsed. I may come back on this in the future but at this time I
think we should stop here until new data is available.

Closes #14